# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for BookBuddy
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  email: string;
  /** @nullable */
  display_name?: string | null;
  daily_goal_minutes: number;
  reminder_enabled: boolean;
  /** @nullable */
  reminder_time?: string | null;
  /** @nullable */
  burnout_window_start?: string | null;
  /** @nullable */
  burnout_window_end?: string | null;
  dark_mode: boolean;
  created_at: string;
}

export interface UpdateProfileRequest {
  /** @nullable */
  display_name?: string | null;
  /**
   * @minimum 1
   * @maximum 600
   */
  daily_goal_minutes?: number;
  reminder_enabled?: boolean;
  /** @nullable */
  reminder_time?: string | null;
  /** @nullable */
  burnout_window_start?: string | null;
  /** @nullable */
  burnout_window_end?: string | null;
  dark_mode?: boolean;
}

export interface Book {
  id: string;
  user_id: string;
  title: string;
  /** @nullable */
  author?: string | null;
  total_pages: number;
  current_page: number;
  is_completed: boolean;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  last_read_at?: string | null;
  is_current: boolean;
  created_at: string;
}

export interface CreateBookRequest {
  title: string;
  /** @nullable */
  author?: string | null;
  /** @minimum 1 */
  total_pages: number;
  /** @minimum 0 */
  current_page?: number;
  is_current?: boolean;
}

export interface UpdateBookRequest {
  title?: string;
  /** @nullable */
  author?: string | null;
  /** @minimum 1 */
  total_pages?: number;
  /** @minimum 0 */
  current_page?: number;
  is_completed?: boolean;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  last_read_at?: string | null;
  is_current?: boolean;
}

export interface ReadingSession {
  id: string;
  user_id: string;
  book_id: string;
  pages_read: number;
  minutes_read: number;
  /** @nullable */
  start_page?: number | null;
  /** @nullable */
  end_page?: number | null;
  session_date: string;
  /** @nullable */
  notes?: string | null;
  created_at: string;
  book?: Book | null;
}

export interface CreateSessionRequest {
  book_id: string;
  /** @minimum 0 */
  pages_read: number;
  /** @minimum 0 */
  minutes_read: number;
  /** @nullable */
  start_page?: number | null;
  /** @nullable */
  end_page?: number | null;
  session_date: string;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardSummary {
  /** Total minutes read today */
  todayMinutes: number;
  /** Number of days with reading activity this week (Mon-Sun) */
  weeklyActiveDays: number;
  /** Current consecutive reading days streak */
  currentStreak: number;
  /** Longest streak ever achieved */
  longestStreak: number;
  /** e.g. '4 of 7 days' */
  weeklyConsistency: string;
  /**
   * Days since the last reading session (null if never read)
   * @nullable
   */
  daysSinceLastSession: number | null;
  /** Total pages read across all sessions */
  totalPagesAllTime: number;
  /** Total minutes read across all sessions */
  totalMinutesAllTime: number;
  /** Total number of reading sessions */
  totalSessionsAllTime: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListSessionsParams = {
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: number;
};