# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for BookBuddy
 * OpenAPI spec version: 0.1.0
 */

export interface Book {
  id: string;
  user_id: string;
  title: string;
  /** @nullable */
  author?: string | null;
  total_pages: number;
  current_page: number;
  is_completed: boolean;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  last_read_at?: string | null;
  is_current: boolean;
  created_at: string;
}