# lib/api-zod/src/generated/types/createBookRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for BookBuddy
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookRequest {
  title: string;
  /** @nullable */
  author?: string | null;
  /** @minimum 1 */
  total_pages: number;
  /** @minimum 0 */
  current_page?: number;
  is_current?: boolean;
}