# lib/api-zod/src/generated/types/createSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for BookBuddy
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSessionRequest {
  book_id: string;
  /** @minimum 0 */
  pages_read: number;
  /** @minimum 0 */
  minutes_read: number;
  /** @nullable */
  start_page?: number | null;
  /** @nullable */
  end_page?: number | null;
  session_date: Date;
  /** @nullable */
  notes?: string | null;
}