# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for BookBuddy
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  /** Total minutes read today */
  todayMinutes: number;
  /** Number of days with reading activity this week (Mon-Sun) */
  weeklyActiveDays: number;
  /** Current consecutive reading days streak */
  currentStreak: number;
  /** Longest streak ever achieved */
  longestStreak: number;
  /** e.g. '4 of 7 days' */
  weeklyConsistency: string;
  /**
   * Days since the last reading session (null if never read)
   * @nullable
   */
  daysSinceLastSession: number | null;
  /** Total pages read across all sessions */
  totalPagesAllTime: number;
  /** Total minutes read across all sessions */
  totalMinutesAllTime: number;
  /** Total number of reading sessions */
  totalSessionsAllTime: number;
}