# lib/api-zod/src/generated/types/readingSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for BookBuddy
 * OpenAPI spec version: 0.1.0
 */
import type { Book } from "./book";

export interface ReadingSession {
  id: string;
  user_id: string;
  book_id: string;
  pages_read: number;
  minutes_read: number;
  /** @nullable */
  start_page?: number | null;
  /** @nullable */
  end_page?: number | null;
  session_date: Date;
  /** @nullable */
  notes?: string | null;
  created_at: string;
  book?: Book | null;
}