# lib/api-zod/src/generated/types/updateBookRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for BookBuddy
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateBookRequest {
  title?: string;
  /** @nullable */
  author?: string | null;
  /** @minimum 1 */
  total_pages?: number;
  /** @minimum 0 */
  current_page?: number;
  is_completed?: boolean;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  last_read_at?: string | null;
  is_current?: boolean;
}