# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for BookBuddy
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileRequest {
  /** @nullable */
  display_name?: string | null;
  /**
   * @minimum 1
   * @maximum 600
   */
  daily_goal_minutes?: number;
  reminder_enabled?: boolean;
  /** @nullable */
  reminder_time?: string | null;
  /** @nullable */
  burnout_window_start?: string | null;
  /** @nullable */
  burnout_window_end?: string | null;
  dark_mode?: boolean;
}